Этап 2-1

+++ main-service/src/main/java/ru/practicum/ewm/category/repository/CategoryRepository.java
@@ -3,5 +3,7 @@ package ru.practicum.ewm.category.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import ru.practicum.ewm.category.model.Category;
 
-public interface CategoryRepository extends JpaRepository<Category, Long> {
+
+public interface CategoryRepository extends JpaRepository<Category, Integer> {
+
 }
